requirements.txt
Django==3.2.8
Pillow==8.4.0
nodeenv
transliterate==1.10.2
factory-boy==3.2.1
django-environ==0.8.1
loguru~=0.5.3
# dev
pre-commit==2.15.0
black==21.9b0
flake8==3.9.2
isort==5.9.3
django-filter==21.1
django_extensions
celery

# test
selenium==4.1.0
